+++ src/main/java/at/minecraftschurli/arsmagicalegacy/AMEventHandler.java
@@ -87,7 +87,6 @@ import net.minecraft.world.level.block.DispenserBlock;
 import net.minecraft.world.level.block.FireBlock;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.entity.LecternBlockEntity;
-import net.minecraft.world.phys.AABB;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.EventBusSubscriber;
@@ -292,12 +291,7 @@ final class AMEventHandler {
         Entity entity = event.getEntity();
         Level level = entity.level();
         if (entity instanceof ItemFrame itemFrame && (itemFrame.hasData(AMAttachments.COMPENDIUM_TIMER) || level.getGameTime() % AMServerConfig.ARCANE_COMPENDIUM_CONVERSION_DURATION.getAsInt() == 0)) {
-            int range = AMServerConfig.ARCANE_COMPENDIUM_CONVERSION_HORIZONTAL_RANGE.getAsInt();
-            BlockPos pos = itemFrame.getPos().offset(itemFrame.getDirection().getNormal().multiply(Math.ceilDiv(range, 2))).below();
-            AMUtil.doCompendiumConversion(itemFrame, level, itemFrame.position(), AABB.encapsulatingFullBlocks(
-                pos.offset(-range / 2, 0, -range / 2),
-                pos.offset(range / 2, 1 - AMServerConfig.ARCANE_COMPENDIUM_CONVERSION_VERTICAL_RANGE.getAsInt(), range / 2)
-            ), itemFrame::getItem, itemFrame::setItem);
+            AMUtil.doCompendiumConversion(itemFrame);
         }
         if (entity instanceof LivingEntity living) {
             ManaHelper manaHelper = ArsMagicaApi.manaHelper();