+++ changelog.md
@@ -15,6 +15,7 @@
 
 - Renamed Liquid Essence to Liquid Etherium
 - Now generates in lakes in plains-like biomes, in addition to the centers of Moonstone meteorites
+- Can now be used to create the Arcane Compendium by placing a book in a nearby item frame, similar to old Ars Magica
 - Now has underwater fog
 - Now can be placed in Cauldrons
 - Liquid Etherium Bucket can now be used by Dispensers